add read

+++ ModuleBase/src/main/java/com/xd/base/file/FileUtils.java
@@ -1,12 +1,18 @@
 package com.xd.base.file;
 
+import android.annotation.SuppressLint;
+import android.content.ContentResolver;
+import android.content.ContentUris;
 import android.content.Context;
 import android.content.Intent;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
+import android.provider.DocumentsContract;
+import android.provider.MediaStore;
 import android.util.Base64;
 import android.util.Log;
 
@@ -14,6 +20,7 @@ import android.util.Log;
 import androidx.core.content.FileProvider;
 
 import com.xd.base.utils.IOUtils;
+import com.xd.base.utils.NStringUtils;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
@@ -46,6 +53,19 @@ import java.util.zip.ZipFile;
 public class FileUtils {
     //采用自己的格式去设置文件，防止文件被系统文件查询到
 
+    //更新 下载APK
+    public static String getFileNameFromPath(String path){
+        if (NStringUtils.isBlank(path)){
+            return null;
+        }
+        String finalPath = null;
+        try {
+            finalPath = path.substring(path.lastIndexOf("/") + 1, path.length());
+        } catch (Exception e) {
+
+        }
+        return finalPath;
+    }
     //获取文件夹
     public static File getFolder(String filePath){
         File file = new File(filePath);
@@ -407,6 +427,8 @@ public class FileUtils {
                     output.flush();
                     output.close();
                     input.close();
+                }else {
+                    copyFolder(temp.getPath(),newPath+File.separator);
                 }
                 //如果游标遇到文件夹
 //                if (temp.isDirectory()) {
@@ -498,11 +520,16 @@ public class FileUtils {
     }
 
 
+
+    public static void unZipFile(String zipFile, String folderPath) {
+        unZipFile(zipFile,folderPath,true);
+    }
+
     /**
      * zipFile 压缩文件
      * folderPath 解压后的文件路径
      * */
-    public static void unZipFile(String zipFile, String folderPath) {
+    public static void unZipFile(String zipFile, String folderPath,boolean deleteZipFile) {
         try {
             File outputFloder = new File(folderPath);
             if(!outputFloder.exists()){
@@ -540,8 +567,10 @@ public class FileUtils {
             e.printStackTrace();
         }
 
-        //解压完成之后删除压缩包
-        deleteFile(zipFile);
+        if (deleteZipFile) {
+            //解压完成之后删除压缩包
+            deleteFile(zipFile);
+        }
     }
 
     /**
@@ -625,7 +654,7 @@ public class FileUtils {
         {
             if(currentFiles[i].isDirectory())//如果当前项为子目录 进行递归
             {
-                copyFile(currentFiles[i].getPath() + "/", toFile + currentFiles[i].getName() + "/");
+                copyFile(currentFiles[i].getPath() + "/", toFile + File.separator + currentFiles[i].getName() + "/");
 
             }else//如果当前项为文件则进行文件拷贝
             {
@@ -725,4 +754,116 @@ public class FileUtils {
             context.startActivity(intent);
         }
     }
+
+
+    @SuppressLint("NewApi")
+    public static String getFilePathByUri(Context context, Uri uri) {
+
+        // 4.4及之后的 是 DocumentProvider。
+        // 比如 content://com.android.providers.media.documents/document/image%3A235700
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT && DocumentsContract.isDocumentUri(context, uri)) {
+
+            // ExternalStorageProvider
+            if (isExternalStorageDocument(uri)) {
+                final String docId = DocumentsContract.getDocumentId(uri);
+                final String[] split = docId.split(":");
+                final String type = split[0];
+                if ("primary".equalsIgnoreCase(type)) {
+                    return Environment.getExternalStorageDirectory() + "/" + split[1];
+                }
+            }
+
+            // DownloadsProvider
+            else if (isDownloadsDocument(uri)) {
+                final String id = DocumentsContract.getDocumentId(uri);
+
+                //处理某些机型（比如Goole Pixel ）ID是raw:/storage/emulated/0/Download/c20f8664da05ab6b4644913048ea8c83.mp4
+                final String[] split = id.split(":");
+                final String type = split[0];
+                if ("raw".equalsIgnoreCase(type)) {
+                    return split[1];
+                }
+
+
+                String[] contentUriPrefixesToTry = new String[]{
+                        "content://downloads/public_downloads",
+                        "content://downloads/my_downloads",
+                        "content://downloads/all_downloads"
+                };
+
+                for (String contentUriPrefix : contentUriPrefixesToTry) {
+                    Uri contentUri = ContentUris.withAppendedId(Uri.parse(contentUriPrefix), Long.valueOf(id));
+                    try {
+                        String path = getDataColumn(context, contentUri, null, null);
+                        if (path != null) {
+                            return path;
+                        }
+                    } catch (Exception e) {
+                    }
+                }
+            }
+
+            // MediaProvider
+            else if (isMediaDocument(uri)) {
+                final String docId = DocumentsContract.getDocumentId(uri);
+                final String[] split = docId.split(":");
+                final String type = split[0];
+                Uri contentUri = null;
+                if ("image".equals(type)) {
+                    contentUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI;
+                } else if ("video".equals(type)) {
+                    contentUri = MediaStore.Video.Media.EXTERNAL_CONTENT_URI;
+                } else if ("audio".equals(type)) {
+                    contentUri = MediaStore.Audio.Media.EXTERNAL_CONTENT_URI;
+                }
+                final String selection = "_id=?";
+                final String[] selectionArgs = new String[]{split[1]};
+                return getDataColumn(context, contentUri, selection, selectionArgs);
+            }
+        }
+
+        // 以 file:// 开头的
+        else if (ContentResolver.SCHEME_FILE.equals(uri.getScheme())) {
+            return uri.getPath();
+        }
+
+        // 以 content:// 开头的，比如 content://media/extenral/images/media/17766
+        // 注意这种情况是一般情况，4.4前后都有
+        else if (ContentResolver.SCHEME_CONTENT.equals(uri.getScheme())) {
+            return getDataColumn(context, uri, null, null);
+        }
+        return null;
+    }
+
+    private static String getDataColumn(Context context, Uri uri, String selection, String[] selectionArgs) {
+        Cursor cursor = null;
+        final String column = MediaStore.Images.Media.DATA;
+        final String[] projection = {column};
+        try {
+            cursor = context.getContentResolver().query(uri, projection, selection, selectionArgs, null);
+            if (cursor != null && cursor.moveToFirst()) {
+                final int column_index = cursor.getColumnIndexOrThrow(column);
+                return cursor.getString(column_index);
+            }
+        }catch (Exception e){
+            e.printStackTrace();
+        } finally {
+            if (cursor != null)
+                cursor.close();
+        }
+        return null;
+    }
+
+    private static boolean isExternalStorageDocument(Uri uri) {
+        return "com.android.externalstorage.documents".equals(uri.getAuthority());
+    }
+
+    private static boolean isDownloadsDocument(Uri uri) {
+        return "com.android.providers.downloads.documents".equals(uri.getAuthority());
+    }
+
+    private static boolean isMediaDocument(Uri uri) {
+        return "com.android.providers.media.documents".equals(uri.getAuthority());
+    }
+
 }
